Please Call Me AdServer in PHP
The assignment is as follows:
Use any language to create a web service that is a very simple
Please-Call-Me ad-server.  The HTTP request must receive the 'from' and
'to' phone numbers which should look something like this:
http://your.server/getAd?from=27821112220&to=27821112221
The server must then use the telephone numbers from the request and the
information in the 2 attached CSV files to return the text of the most
most appropriate Ad. For example, I'll want to see something like this
in the response:
Please call John on +27 82 111 2220 ** Dial *120*123443#  to get your
free coupons at Builders Warehouse **